refactor: centralize test stubs for quart and cv2
