+++ tests/test_telegram_notify_thread_stop.py
@@ -5,10 +5,9 @@ from pathlib import Path
 # เพิ่ม path ของ src เพื่อให้สามารถ import packages ได้
 sys.path.insert(0, str(Path(__file__).resolve().parents[1] / "src"))
 
-# stub cv2 module so import works without opencv
-cv2_stub = types.ModuleType("cv2")
-cv2_stub.IMWRITE_JPEG_QUALITY = 1
-sys.modules["cv2"] = cv2_stub
+from tests.stubs import stub_cv2
+
+cv2_stub = stub_cv2()
 
 # stub requests module and its submodules
 requests_stub = types.ModuleType("requests")
